requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
channels==4.3.0
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
distlib==0.3.9
Django==5.2.4
django-auditlog==3.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
django-user-agents==0.4.0
djangorestframework==3.16.0
drf-jwt==1.19.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.18.0
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
minio==7.2.16
msgpack==1.1.1
nodeenv==1.9.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
platformdirs==4.3.8
pre_commit==4.2.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
virtualenv==20.31.2
zope.interface==7.2
